+++ src/entities/environment/clouds/index.ts
@@ -1,0 +1,1 @@
+export * from "./clouds";
