+++ src/entities/environment/grass/index.ts
@@ -1,0 +1,2 @@
+export * from "./grass";
+export * from "./grassOptions";
