+++ src/entities/environment/ground/index.ts
@@ -1,0 +1,1 @@
+export * from "./ground";
