+++ src/entities/environment/rocks/index.ts
@@ -1,0 +1,2 @@
+export * from "./rocks";
+export * from "./rockOptions";
