+++ src/entities/environment/skyBox/index.ts
@@ -1,0 +1,2 @@
+export * from "./skyBox";
+export * from "./skyBoxOptions";
