+++ src/entities/environment/tree/index.ts
@@ -1,0 +1,6 @@
+export * from "./branch";
+export * from "./enums";
+export * from "./rng";
+export * from "./textures";
+export * from "./tree";
+export * from "./treeOptions";
